hw: UART Driver for AM65x and Dummy Clock Creation
Adding AM65x UART Driver along with the existing OMAP UART

This process required creation of a dummy clock outside of omap_clk.c and thus
the clk structure has been redefined outside of omap_clk.c to omap.h. This is
done so as to avoid use of omap_findclk() function which requires creation of
an OMAP MPU State which we have not created in our Machine.

Signed-off-by: Aman Jain <amanajain98@gmail.com>