hw: arm: AM65x Board
Implementation of the TI AM65x Model inside QEMU with minimal SoC
identification. The TI AM65x SoC is a part of the K3 SoC Architecture
platform.It aims to meet the Complex processing needs of modern
embedded products.

Peripherals implemented so far are the A53, GICv3
and the UART.

Signed-off-by: Aman Jain <amanajain98@gmail.com>